implements arbitrary camera dimension with arbitrary camera cell size
we can now have camera cells bigger than world cells and it works (less
than perfectly tested though).

MAJOR CAVEAT: right now, to make things work, the printer is allowed to
go to the edge of the build platform, in which case, the camera would be
grabbing out of bounds cells from the grid. to keep this from making
everything explode, if the location we're asking for is out of bounds
we just say there is no material there. this is not the right behavior.

will implement functionality to keep the printer from moving off the
build platform. however, this is going to require some nontrivial work
on what contains what in the simulation. needed to commit these changes
before working on that stuff